Fixed incorrect mapping for floats by implementing function float mapf(float, float, float, foat, float)

+++ light_sensor_six_lego/printHelpers.ino
@@ -39,7 +39,7 @@ void printPolePosition() {
 void printMessage() {
   Serial.print("message:\t");
   for (int i = 0; i < 8; i++) {
-    Serial.print(message[i]);
+    Serial.print((int8_t)message[i]);
     Serial.print("\t");
   }
   Serial.println();
